[fix:tars plugin #3406] tars using shared thread pool
